update charts folder structure
- moves chart header into chart component
- pulls out highcharts chart into seperate component
